Fixes for OpenWatcom:   regex.c(3817): Warning! W400: Array subscript is of type plain char   smartlist.c(363): Error! E1071: Type of parameter 4 does not agree with previous definition   I.e. the compare-function for 'qsort()' do NOT need to be '__cdecl'.

+++ src/regex.c
@@ -55,7 +55,7 @@ typedef unsigned long int active_reg_t;
 /* How many characters in the character set.
  */
 #define CHAR_SET_SIZE 256
-#define SYNTAX(c)     re_syntax_table[c]
+#define SYNTAX(c)     re_syntax_table [(int)(c)]
 
 static char re_syntax_table [CHAR_SET_SIZE];
